refactor: extract shell-process toolkit policy boundary
