+++ src/core/tools/README.md
@@ -15,7 +15,7 @@ workspace.
 - Knowledge and memory-surface tools under `toolkits/knowledge/`.
 - External context tools under `toolkits/external-context/`.
 - Internal workflow/state tools under `toolkits/internal/`.
-- Shell inspect/mutate tools and their current command policy rules.
+- Shell/process tools and their command policy rules under `toolkits/shell-process/`.
 
 ## Does Not Own
 
@@ -36,15 +36,11 @@ workspace.
 - `toolkits/external-context/`: provider-backed web search and image inspection
   tools plus their toolkit composition.
 - `toolkits/internal/`: internal structured workflow tools such as
-  `update_plan`, plus the current M12 runtime-owned shell/process grouping used
-  to preserve default bundle behavior.
-- `toolkits/internal/run-shell.ts`: current shell/process tool implementation
-  and shell policy location.
+  `update_plan`.
+- `toolkits/shell-process/`: shell inspect/mutate execution, command policy,
+  and toolkit composition.
 - `toolkits/*/toolkit.ts`: production toolkit composition entry points used by
   runtime default-tool assembly.
-- Shell/process is temporarily housed under `toolkits/internal/` for M12
-  compatibility. M13 is the planned milestone for clearer shell/process policy
-  extraction.
 
 ## Public Entry Points
 
@@ -54,8 +50,9 @@ workspace.
 - `toolkits/coding-files/*`: coding file tools.
 - `toolkits/knowledge/*`: knowledge and memory-surface tools.
 - `toolkits/external-context/*`: web and image tools.
-- `toolkits/internal/*`: internal workflow/state tools, including the current
-  M12 shell/process implementation at `toolkits/internal/run-shell.ts`.
+- `toolkits/internal/*`: internal workflow/state tools.
+- `toolkits/shell-process/*`: shell inspect/mutate tools, shell policy, and
+  toolkit entry points.
 
 ## Extension Points
 
@@ -80,10 +77,11 @@ workspace.
   `toolkits/external-context/`.
 - To add an internal structured workflow tool, place it under
   `toolkits/internal/`.
-- To change shell/process tool behavior or shell classification policy, update
-  `toolkits/internal/run-shell.ts` for the current M12 structure, plus the
-  classification and approval-rule compatibility tests. M13 is the planned
-  milestone for extracting shell/process policy further.
+- To change shell/process tool behavior, update
+  `toolkits/shell-process/run-shell.ts`.
+- To change shell classification or policy, update
+  `toolkits/shell-process/shell-policy.ts` plus the classification and
+  approval-rule compatibility tests.
 - To change a tool output shape, update trace/review projection tests if hosts
   depend on that output.
