+++ src/core/tools/toolkits/shell-process/index.ts
@@ -1,0 +1,18 @@
+export { shellProcessToolkit } from './toolkit.js';
+export {
+  classifyShellCommandPolicy,
+  createRunShellInspectTool,
+  createRunShellMutateTool,
+  createRunShellTool,
+  DEFAULT_INSPECT_RULES,
+  DEFAULT_MUTATE_RULES,
+  runShellCommand,
+} from './run-shell.js';
+export type {
+  RunShellCapability,
+  RunShellOptions,
+  RunShellPolicyDecision,
+  RunShellRisk,
+  RunShellRule,
+  RunShellScope,
+} from './run-shell.js';
